Deleted Duplicated class and merged it with Original class. I have not tested if everything still works
